requirements.txt
greenlet==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
